refactor(payment): Phase 2a move integration event payloads to Contracts/
- Move 6 payload records (PaymentAuthorized/Failed/Captured/Refunded/
  Voided, consumed OrderCreated) to Payment.Service.Contracts.Integration.
- Delete IntegrationEvents/Events/; EventHandlers stay until Phases 7c/7d.
