+++ payment-microservice/Payment.Service/Contracts/Integration/PaymentAuthorizedEvent.cs
@@ -1,6 +1,6 @@
 using ECommerce.Shared.Infrastructure.EventBus;
 
-namespace Payment.Service.IntegrationEvents.Events;
+namespace Payment.Service.Contracts.Integration;
 
 public record PaymentAuthorizedEvent(
     Guid PaymentId,
