+++ payment-microservice/Payment.Service/IntegrationEvents/EventHandlers/AuthorizePaymentCommandHandler.cs
@@ -3,9 +3,9 @@ using ECommerce.Shared.Infrastructure.EventBus;
 using ECommerce.Shared.Infrastructure.EventBus.Abstractions;
 using ECommerce.Shared.Infrastructure.Outbox;
 using ECommerce.Shared.IntegrationEvents.Commands;
+using Payment.Service.Contracts.Integration;
 using Payment.Service.Infrastructure.Data;
 using Payment.Service.Infrastructure.Gateways;
-using Payment.Service.IntegrationEvents.Events;
 using Payment.Service.Models;
 using Payment.Service.Observability;
