+++ payment-microservice/Payment.Tests/Api/InternalOutboxEndpointsTests.cs
@@ -3,7 +3,7 @@ using System.Net.Http.Json;
 using System.Text.Json;
 using ECommerce.Shared.Infrastructure.Outbox;
 using Microsoft.Extensions.DependencyInjection;
-using Payment.Service.IntegrationEvents.Events;
+using Payment.Service.Contracts.Integration;
 
 namespace Payment.Tests.Api;
